+++ examples/multithread.rs
@@ -22,7 +22,8 @@ children:
       - name: snippet
         base_path: a.result__snippet
         extract: html
-        pipeline: [ [ policy_highlight ], [ trim_space ] ]
+        sanitize_policy: highlight
+        pipeline: [ [ trim_space ] ]
 ";
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {